+++ planning/obstacle_avoidance_planner/include/obstacle_avoidance_planner/mpt_optimizer.hpp
@@ -46,6 +46,7 @@
 #include "obstacle_avoidance_planner/common_structs.hpp"
 #include "obstacle_avoidance_planner/vehicle_model/vehicle_model_interface.hpp"
 #include "osqp_interface/osqp_interface.hpp"
+#include "tier4_autoware_utils/system/stop_watch.hpp"
 #include "tier4_autoware_utils/tier4_autoware_utils.hpp"
 
 #include "autoware_auto_planning_msgs/msg/path_point.hpp"
